+++ lib/hexapdf/type/acro_form.rb
@@ -47,6 +47,7 @@ module HexaPDF
       autoload(:VariableTextField, 'hexapdf/type/acro_form/variable_text_field')
       autoload(:TextField, 'hexapdf/type/acro_form/text_field')
       autoload(:ButtonField, 'hexapdf/type/acro_form/button_field')
+      autoload(:ChoiceField, 'hexapdf/type/acro_form/choice_field')
 
       autoload(:AppearanceGenerator, 'hexapdf/type/acro_form/appearance_generator')