+++ lib/hexapdf/type/acro_form/form.rb
@@ -173,6 +173,22 @@ module HexaPDF
           create_field(name, :Btn).tap(&:initialize_as_radio_button)
         end
 
+        # Creates a combo box with the given name and adds it to the form.
+        #
+        # The +name+ may contain dots to signify a field hierarchy. If so, the referenced parent
+        # fields must already exist. If it doesn't contain dots, a top-level field is created.
+        def create_combo_box(name)
+          create_field(name, :Ch).tap(&:initialize_as_combo_box)
+        end
+
+        # Creates a list box with the given name and adds it to the form.
+        #
+        # The +name+ may contain dots to signify a field hierarchy. If so, the referenced parent
+        # fields must already exist. If it doesn't contain dots, a top-level field is created.
+        def create_list_box(name)
+          create_field(name, :Ch).tap(&:initialize_as_list_box)
+        end
+
         # Returns the dictionary containing the default resources for form field appearance streams.
         def default_resources
           self[:DR] ||= document.wrap({}, type: :XXResources)